move username resignation components
