+++ src/domains/transaction/components/SendUsernameResignationSidePanel/ReviewStep.tsx
