+++ src/domains/transaction/components/SendUsernameResignationSidePanel/SendUsernameResignationSidePanel.tsx
@@ -4,8 +4,8 @@ import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import cn from "classnames";
 import { useTranslation } from "react-i18next";
-import { FormStep } from "@/domains/transaction/pages/SendUsernameResignation/FormStep";
-import { ReviewStep } from "@/domains/transaction/pages/SendUsernameResignation/ReviewStep";
+import { FormStep } from "@/domains/transaction/components/SendUsernameResignationSidePanel/FormStep";
+import { ReviewStep } from "@/domains/transaction/components/SendUsernameResignationSidePanel/ReviewStep";
 import { usePendingTransactions } from "@/domains/transaction/hooks/use-pending-transactions";
 import { Form } from "@/app/components/Form";
 import { TabPanel, Tabs } from "@/app/components/Tabs";
