+++ src/domains/transaction/components/SendUsernameResignationSidePanel/index.ts
@@ -1,0 +1,2 @@
+export * from "./FormStep";
+export * from "./ReviewStep";
