--- src/domains/transaction/pages/SendUsernameResignation/index.ts
@@ -1,3 +1,0 @@
-export * from "./FormStep";
-export * from "./ReviewStep";
-export { SendUsernameResignation as default } from "./SendUsernameResignation";
